Moved all the code for initializing astronaut and mission from Main() program thread to Mission Control initialization. Now Mission Control monitors astronaut's vitals every second. If any of the suit vitals are outside the expected range, mission control warns astronaut. This is a better approach since Unity UI will have it's own thread running and Mission Control class can monitor suit vitals (and mission, navigation, etc) with multiple UI independent threads.
